add: Timelog table, BudgetedHours, estimated hours

+++ TaskTrackingSystem.Shared/Models/Report/TaskReportDto.cs
@@ -16,7 +16,7 @@ namespace TaskTrackingSystem.Shared.Models.Report
         public string? AssignedToUser { get; set; }
         public long? AssignedToUserId { get; set; }
         public string? AssignedByUser { get; set; }
-        public int? EstimatedHours { get; set; }
+        public decimal? EstimatedHours { get; set; }
         public DateTime DueDate { get; set; }
         public DateTime CreatedAt { get; set; }
     }
